reworked ASGIAuthenticationMiddleware to be a subclass of ASGIMiddleware
