+++ litestar/middleware/__init__.py
@@ -1,6 +1,5 @@
 from litestar.middleware.authentication import (
     AbstractAuthenticationMiddleware,
-    ASGIAuthenticationMiddleware,
     AuthenticationResult,
 )
 from litestar.middleware.base import (
@@ -11,7 +10,6 @@ from litestar.middleware.base import (
 )
 
 __all__ = (
-    "ASGIAuthenticationMiddleware",
     "ASGIMiddleware",
     "AbstractAuthenticationMiddleware",
     "AbstractMiddleware",
