+++ litestar/middleware/base.py
@@ -4,7 +4,8 @@ import abc
 from abc import abstractmethod
 from typing import TYPE_CHECKING, Any, Callable, Protocol, runtime_checkable
 
-from litestar.enums import ScopeType
+from litestar.connection import ASGIConnection
+from litestar.enums import HttpMethod, ScopeType
 from litestar.middleware._utils import (
     build_exclude_path_pattern,
     should_bypass_middleware,
@@ -20,8 +21,10 @@ __all__ = (
 
 
 if TYPE_CHECKING:
-    from litestar.types import Scopes
-    from litestar.types.asgi_types import ASGIApp, Receive, Scope, Send
+    from collections.abc import Sequence
+
+    from litestar.middleware import AuthenticationResult
+    from litestar.types import ASGIApp, Method, Receive, Scope, Scopes, Send
 
 
 @runtime_checkable
@@ -249,3 +252,52 @@ class ASGIMiddleware(abc.ABC):
         """
 
         raise NotImplementedError
+
+
+class ASGIAuthenticationMiddleware(ASGIMiddleware):
+    """ASGI Authentication Middleware that allows users to create their own authentication middleware by extending it
+    and overriding :meth:`ASGIAuthenticationMiddleware.authenticate_request`.
+    """
+
+    scopes: tuple[ScopeType, ...] = (ScopeType.HTTP, ScopeType.WEBSOCKET, ScopeType.ASGI)
+    exclude_path_pattern: str | tuple[str, ...] | None = None
+    exclude_opt_key: str = "exclude_from_auth"
+    exclude_http_methods: Sequence[Method] = (HttpMethod.OPTIONS,)
+
+    async def handle(self, scope: Scope, receive: Receive, send: Send, next_app: ASGIApp) -> None:
+        """Create the actual middleware callable"""
+        authenticate_request = self.authenticate_request
+        exclude_pattern = build_exclude_path_pattern(exclude=self.exclude_path_pattern, middleware_cls=type(self))
+        scopes = set(self.scopes)
+        exclude_opt_key = self.exclude_opt_key
+        exclude_http_methods = self.exclude_http_methods
+
+        if not should_bypass_middleware(
+            exclude_http_methods=exclude_http_methods,
+            exclude_opt_key=exclude_opt_key,
+            exclude_path_pattern=exclude_pattern,
+            scope=scope,
+            scopes=scopes,  # type: ignore[arg-type]
+        ):
+            auth_result = await authenticate_request(ASGIConnection(scope))
+            scope["user"] = auth_result.user
+            scope["auth"] = auth_result.auth
+        await next_app(scope, receive, send)
+
+    @abstractmethod
+    async def authenticate_request(self, connection: ASGIConnection) -> AuthenticationResult:
+        """Receive the ASGI connection and return an :class:`AuthenticationResult`.
+
+        Notes:
+            - This method must be overridden by subclasses.
+
+        Args:
+            connection: An :class:`ASGIConnection <litestar.connection.ASGIConnection>` instance.
+
+        Raises:
+            NotAuthorizedException | PermissionDeniedException: if authentication fails.
+
+        Returns:
+            An instance of :class:`AuthenticationResult <litestar.middleware.authentication.AuthenticationResult>`.
+        """
+        raise NotImplementedError("handle must be overridden by subclasses")
